Set user null when request by an expired token

+++ frontend/src/reducers/categoryReducer.js
@@ -1,5 +1,7 @@
 import { createSlice } from "@reduxjs/toolkit";
 import categoryService from "../services/category";
+import { getToken, isTokenExpired } from "../services/login";
+import { rmUserFn } from "./userReducer";
 
 const initialState = [];
 
@@ -32,6 +34,11 @@ export const { createCategory, setCategories, dltCategory, editCategory } =
 
 export const createNewCategory = (category) => {
     return async (dispatch) => {
+        const token = getToken();
+        if (isTokenExpired(token)) {
+            dispatch(rmUserFn());
+            return;
+        }
         const newCategory = await categoryService.addCategory(category);
         dispatch(createCategory(newCategory));
     };
@@ -39,6 +46,11 @@ export const createNewCategory = (category) => {
 
 export const setAllCategories = () => {
     return async (dispatch) => {
+        const token = getToken();
+        if (isTokenExpired(token)) {
+            dispatch(rmUserFn());
+            return;
+        }
         const categories = await categoryService.getAll();
         dispatch(setCategories(categories));
     };
@@ -46,6 +58,11 @@ export const setAllCategories = () => {
 
 export const updateCategory = (category) => {
     return async (dispatch) => {
+        const token = getToken();
+        if (isTokenExpired(token)) {
+            dispatch(rmUserFn());
+            return;
+        }
         await categoryService.editCategory(category);
         dispatch(editCategory(category));
     };
@@ -53,6 +70,11 @@ export const updateCategory = (category) => {
 
 export const deleteCategory = (cateId) => {
     return async (dispatch) => {
+        const token = getToken();
+        if (isTokenExpired(token)) {
+            dispatch(rmUserFn());
+            return;
+        }
         await categoryService.deleteCategory(cateId);
         dispatch(dltCategory(cateId));
     };
